Allow specification of custom classifier paths.
This appears to be necessary in the latest version as it does not
find the built-in ones any more.  Further, there is no way to
configure them in the INI file anyway.  With this patch, a custom
classifier JSON file can be placed next to the custom configuration
INI file.